feat: :lock: Add Auth Policy
