+++ src/apps/App.Web.Client/Views/Account/AccessDenied.cshtml
@@ -1,0 +1,19 @@
+﻿@{
+    ViewData["Title"] = "Access Denied";
+    var code = ViewBag.StatusCode ?? 403;
+    var message = ViewBag.Message as string;
+}
+    <section class="content d-flex align-items-center justify-content-center" style="min-height: 80vh;">
+        <div class="error-page text-center">
+            <h2 class="headline text-danger">@code</h2>
+
+            <div class="error-content">
+                <h3><i class="fas fa-exclamation-triangle text-danger"></i> Sorry, you do not have permission to view this page.</h3>
+
+                <p>
+                    @Html.Raw(message) <br />
+                    You may <a href="/">return to dashboard</a> or contact your administrator.
+                </p>
+            </div>
+        </div>
+    </section>
