+++ src/apps/App.Web.Client/Views/Home/Index.cshtml
@@ -8,19 +8,19 @@
     <div class="app-content-header">
         <div class="container-fluid">
             <div class="row">
-                <div class="col-sm-6"><h3 class="mb-0">Dashboard</h3></div>
+                <div class="col-sm-6"><h3 class="mb-0">Welcome!</h3></div>
             </div>
         </div>
     </div>
     <br/>
     <div class="app-content">
         <div class="container-fluid">
-            <div class="row">
+            @* <div class="row">
                 @await Component.InvokeAsync("InfoBoxLoading", InfoBoxEnum.TotalBalance.ToString())
                 @await Component.InvokeAsync("InfoBoxLoading", InfoBoxEnum.MonthlyIncome.ToString())
                 @await Component.InvokeAsync("InfoBoxLoading", InfoBoxEnum.MonthlyExpense.ToString())
                 @await Component.InvokeAsync("InfoBoxLoading", InfoBoxEnum.SavingsRate.ToString())
-            </div>
+            </div> *@
     </div>
 </div>
